requirements.txt
sentence-transformers>=2.2.0
torch>=1.9.0
numpy>=1.21.0
python-dotenv
openai

